+++ R/disclosure-helpers.R
@@ -627,6 +627,7 @@ app_guardrail_server <- function(id, state, app_refuse = .app_refuse) {
 
     shiny::observeEvent(input$abort_flagged, ignoreNULL = TRUE, {
       shiny::removeModal()
+      state$upload_source <- NULL
       state$raw_data <- NULL
       state$profile <- NULL
       state$roles <- NULL